% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_hbar.R
\name{chart_hbar}
\alias{chart_hbar}
\title{create a horizontal barplot}
\usage{
chart_hbar(ano, n)
}
\arguments{
\item{ano}{years}

\item{n}{variable}
}
\value{
Return a graph
}
\description{
create a horizontal barplot
}
