% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_histogram.R
\name{chart_histogram}
\alias{chart_histogram}
\title{create histograms}
\usage{
chart_histogram(variable, data, ano_base_alter)
}
\arguments{
\item{variable}{indicators}

\item{data}{data set}

\item{ano_base_alter}{year start data}
}
\value{
Return a graph
}
\description{
A short description...
}
