% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_line.R
\name{chart_line}
\alias{chart_line}
\title{Chart Line}
\usage{
chart_line(data, variable)
}
\arguments{
\item{data}{data set}

\item{variable}{variable to plot}
}
\value{
Return a graph
}
\description{
Chart Line
}
