% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_first_dif.R
\name{compare_first_dif}
\alias{compare_first_dif}
\title{Compare the first difference}
\usage{
compare_first_dif(var)
}
\arguments{
\item{var}{variable or indicator}
}
\value{
check
}
\description{
A short description...
}
