% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_lag.R
\name{compare_lag}
\alias{compare_lag}
\title{Compare values}
\usage{
compare_lag(var)
}
\arguments{
\item{variable}{indicator or variable}
}
\value{
Vector with TRUE or FALSE
}
\description{
Compare the actual value with the before
}
