% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conferir_indicadores_longo.R
\name{conferir_indicadores_longo}
\alias{conferir_indicadores_longo}
\title{conferir_indicadores_longo}
\usage{
conferir_indicadores_longo(nome_colunas)
}
\arguments{
\item{nome_colunas}{Names of indicators}
}
\value{
A vector with incorrect indicators names.
}
\description{
nome deve conter todos os indicadores na coluna "indicador".
A funcao confere os nomes dos indicadores e aponta qual deve ser corrigido.
}
