% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conferir_nomes_indicadores.R
\name{conferir_nomes_indicadores}
\alias{conferir_nomes_indicadores}
\title{conferir_nomes_indicadores}
\usage{
conferir_nomes_indicadores(nome_colunas)
}
\arguments{
\item{nome_colunas}{Names of indicators}
}
\value{
A vector with correct indicators names.
}
\description{
nome_colunas deve conter todos os indicadores ( nemso : "CHAVE","IBGE6","IBGE7","ANO").
A funcao conferir_nomes_indicadores retorna um vetor com os nomes das colunas corrigidas.
}
