% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_output.R
\name{create_output}
\alias{create_output}
\title{Create Report}
\usage{
create_output(
  mydir = getwd(),
  data_path = NULL,
  lista_cidades = NULL,
  fator_mediana = 0.2,
  regras = NULL
)
}
\arguments{
\item{mydir}{Output directory where are the files}

\item{data_path}{Path to data set}

\item{lista_cidades}{List of cities codes with 7 digits. If NULL, will run with 6 cities in the metropolitan area.}

\item{fator_mediana}{Factor applied to the moving median. The result is a threshold with k\% above or below the median. Default is 20\%.}

\item{regras}{Selects a set of variables/indicators related to the corresponding distributional check/rules. If \code{regras = NULL}, apply all rules to all variables. If the object is set to \code{regras = list(outlier = c("EX_INDICATOR1", "EX_INDICATOR2"), min = "all", max = c("EX_INDICATOR4", "EX_INDICATOR12")  )}, the selected rules are applied to the chosen variables/indicators. One or more indicators/variables may be chosen for the rules. This checking mostly refers to the procedure of detecting the outliers or non-standard observations.}
}
\value{
Report
}
\description{
This function return the report with all statistical analysis and critial rules
}
