% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_report.R
\name{create_report}
\alias{create_report}
\title{Create Report}
\usage{
create_report()
}
\arguments{
\item{path_data}{Path to dataset}
}
\value{
Report
}
\description{
This function return the report with all statistical analysis and critial rules
}
