% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_reporting.R
\name{create_reporting}
\alias{create_reporting}
\title{Create Report}
\usage{
create_reporting(path_data, path_metadata)
}
\arguments{
\item{path_data}{Path to dataset}
}
\value{
Report
}
\description{
This function return the report with all statistical analysis and critial rules
}
