% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creating_report.R
\name{creating_report}
\alias{creating_report}
\title{Create Report}
\usage{
creating_report(mydir = getwd())
}
\arguments{
\item{path_data}{Path to dataset}

\item{knitmydir}{output knit}
}
\value{
Report
}
\description{
This function return the report with all statistical analysis and critial rules
}
