% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_valid.R
\name{data_valid}
\alias{data_valid}
\title{summary_aux}
\usage{
data_valid(df, regras, variables, cidades)
}
\arguments{
\item{df}{A list with rules applied to variables.}

\item{variables}{A vector with names of variables/indicators.}

\item{regra}{The name of rule to filter data in df.}
}
\value{
A data frame with summaries of results.
}
\description{
A support function used to summaries rules in data frame.
}
