% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptiveanalisys.R
\name{descriptiveanalisys}
\alias{descriptiveanalisys}
\title{descriptiveanalisys}
\usage{
descriptiveanalisys(variable, data)
}
\arguments{
\item{variable}{Variable to be analysis}

\item{data}{data set containing variable}
}
\value{
statistics of the data
}
\description{
function for calculate the descriptive statistics of the data
}
