% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_files.R
\name{get_files}
\alias{get_files}
\title{get output files}
\usage{
get_files(output_path)
}
\arguments{
\item{output_path}{Path representing the working directory}
}
\value{
Return a file at output
}
\description{
get output files
}
