% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hampel_filter.R
\name{hampel_filter}
\alias{hampel_filter}
\title{Outlier data filter}
\usage{
hampel_filter(variable)
}
\arguments{
\item{variable}{variable or indicators}
}
\value{
result
}
\description{
A short description...
}
