% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_year_variable.R
\name{remove_year_variable}
\alias{remove_year_variable}
\title{Remove years from variable}
\usage{
remove_year_variable(data, variable, exclusions_years)
}
\arguments{
\item{data}{Dataset}

\item{variable}{Variable to remove the years}

\item{exclusions_years}{List of deleted years}
}
\value{
A dataset without years selected
}
\description{
Remove years from variable
}
