% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_aux.R
\name{summary_aux}
\alias{summary_aux}
\title{summary_aux}
\usage{
summary_aux(df, regra, variables)
}
\arguments{
\item{df}{A list with rules applied to variables.}

\item{regra}{The name of rule to filter data in df.}

\item{variables}{A vector with names of variables/indicators.}
}
\value{
A data frame with summaries of results.
}
\description{
A support function used to summaries rules in data frame.
}
