% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_descriptive.R
\name{table_descriptive}
\alias{table_descriptive}
\title{create descriptive table}
\usage{
table_descriptive(variable, data, ano_base_alter)
}
\arguments{
\item{variable}{variable or indicator}

\item{data}{Dataset}

\item{ano_base_alter}{start year}
}
\value{
A dataframe with de summary statistics
}
\description{
A short description...
}
