% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_quantile.R
\name{table_quantile}
\alias{table_quantile}
\title{Quantile table}
\usage{
table_quantile(variable, data)
}
\arguments{
\item{variable}{variable or indicator}

\item{data}{data set}
}
\value{
result
}
\description{
A short description...
}
