% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_sum.R
\name{total_sum}
\alias{total_sum}
\title{Calculate the total of variable to compare}
\usage{
total_sum(variable, total = 100)
}
\arguments{
\item{total}{}
}
\description{
Calculate the total of variable to compare
}
