% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_boxplot.R
\name{chart_boxplot}
\alias{chart_boxplot}
\title{create a boxplot}
\usage{
chart_boxplot(variable, data, ano_base_alter)
}
\arguments{
\item{ano_base_alter}{year start data}
}
\value{
result
}
\description{
A short description...
}
