% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_output.R
\name{create_output}
\alias{create_output}
\title{Create Report}
\usage{
create_output(mydir = getwd(), data_path = NULL)
}
\arguments{
\item{mydir}{Output directory where are the files}

\item{data_path}{Path to data set}
}
\value{
Report
}
\description{
This function return the report with all statistical analysis and critial rules
}
