% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_output.R
\name{create_output}
\alias{create_output}
\title{Create Report}
\usage{
create_output(
  mydir = getwd(),
  knitmydir = getwd(),
  path_file_rules,
  categorical_variable,
  deleted_variable,
  begin_years,
  deleted_years = c(`NA` = NA_integer_)
)
}
\arguments{
\item{mydir}{Output directory}

\item{knitmydir}{output knit}

\item{categorical_variable}{List of qualitative variables}

\item{deleted_variable}{List of deleted variables}

\item{path_data}{Path to dataset}
}
\value{
Report
}
\description{
This function return the report with all statistical analysis and critial rules
}
