% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difpercentual.R
\name{difpercentual}
\alias{difpercentual}
\title{Diference in percentual}
\usage{
difpercentual(df, variable)
}
\arguments{
\item{variable}{}
}
\value{
Data frame containing the difference
}
\description{
Calculate the difference between the percentage of municipal participation in the year and the participation in previous year
}
