% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_function.R
\name{outlier_function}
\alias{outlier_function}
\title{Detect outlier}
\usage{
outlier_function(variable)
}
\arguments{
\item{variable}{quantitative variable}
}
\value{
Vector with TRUE or FALSE
}
\description{
Detect outlier as well as dataMaid::identifyOutliers()
'mc' is the 'medcouple', a robust concept and estimator of skewness, and a and b are appropriate constants (-4 and 3).
The medcouple is defined as a scaled median difference of the left and right half of distribution,
and hence not based on the third moment as the classical skewness.
}
