% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_function.R
\name{outlier_function}
\alias{outlier_function}
\title{Detect outlier}
\usage{
outlier_function(variable, a = -4, b = 3)
}
\arguments{
\item{variable}{quantitative variable}
}
\value{
Vector with TRUE or FALSE
}
\description{
Detect outlier as well as dataMaid::identifyOutliers()
}
