% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_regras.R
\name{parse_regras}
\alias{parse_regras}
\title{parse_regras}
\usage{
parse_regras(regras_por_var, todas_vars)
}
\arguments{
\item{regras_por_var}{Vector or list of rules with corresponding variables. Eg:regras_por_var <- "all" or regras_por_var <- c("outlier" = "all", "min" = "pib,populacao,area") or regras_por_var <- list(outlier = c("pib", "populacao"), min = "all", max = c("idh", "renda_per_capita")  )}

\item{todas_vars}{Vector with all variables in data base}
}
\value{
List with mapped variables and rules.
}
\description{
Support function used to map specific rules to correspondent variables.
}
