% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_median.R
\name{roll_median}
\alias{roll_median}
\title{roll_median}
\usage{
roll_median(variable, fator)
}
\arguments{
\item{variable}{A variable representing a series of observations.}

\item{fator}{A scale parameter.}
}
\value{
Vector with TRUE or FALSE
}
\description{
This function computes the roll median of ordered observations.
}
