% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_descriptive.R
\name{table_descriptive}
\alias{table_descriptive}
\title{create descriptive table}
\usage{
table_descriptive(variable, data)
}
\arguments{
\item{data}{Dataset}
}
\value{
A dataframe with de summary statistics
}
\description{
A short description...
}
